grafici da fare bene

+++ cuda/kernel_cuda_0.cu
@@ -6,12 +6,12 @@
 // =========================================================================
 // 1. IL KERNEL GPU
 // =========================================================================
-__global__ void gemm_cuda_kernel(int M, int N, int k, const double *A, const double *X, double *Y) {
+__global__ void gemm_cuda_kernel(int M, int N, int k, const float *A, const float *X, float *Y) {
     int i = blockIdx.x * blockDim.x + threadIdx.x;
 
     if (i < M) {
         for (int j = 0; j < N; j++) {
-            double a_val = A[i * N + j];
+            float a_val = A[i * N + j];
             for (int p = 0; p < k; p++) {
                 Y[i * k + p] += a_val * X[j * k + p];
             }
@@ -22,14 +22,14 @@ __global__ void gemm_cuda_kernel(int M, int N, int k, const double *A, const dou
 // =========================================================================
 // 2. L'INTERFACCIA CPU-GPU
 // =========================================================================
-static double *d_A = NULL;
-static double *d_X = NULL;
-static double *d_Y = NULL;
+static float *d_A = NULL;
+static float *d_X = NULL;
+static float *d_Y = NULL;
 
-void setup_device_memory(int M, int N, int k, const double *A, const double *X) {
-    size_t size_A = (size_t)M * N * sizeof(double);
-    size_t size_X = (size_t)N * k * sizeof(double);
-    size_t size_Y = (size_t)M * k * sizeof(double);
+void setup_device_memory(int M, int N, int k, const float *A, const float *X) {
+    size_t size_A = (size_t)M * N * sizeof(float);
+    size_t size_X = (size_t)N * k * sizeof(float);
+    size_t size_Y = (size_t)M * k * sizeof(float);
 
     // Allocazione e Copia Protette
     CHECK_CUDA(cudaMalloc((void**)&d_A, size_A));
@@ -40,8 +40,8 @@ void setup_device_memory(int M, int N, int k, const double *A, const double *X) 
     CHECK_CUDA(cudaMemcpy(d_X, X, size_X, cudaMemcpyHostToDevice));
 }
 
-void compute_local_gemm(int M, int N, int k, const double *A, const double *X, double *Y) {
-    CHECK_CUDA(cudaMemset(d_Y, 0, (size_t)M * k * sizeof(double)));
+void compute_local_gemm(int M, int N, int k, const float *A, const float *X, float *Y) {
+    CHECK_CUDA(cudaMemset(d_Y, 0, (size_t)M * k * sizeof(float)));
 
     int threadsPerBlock = 256;
     int blocksPerGrid = (M + threadsPerBlock - 1) / threadsPerBlock;
@@ -57,9 +57,9 @@ void compute_local_gemm(int M, int N, int k, const double *A, const double *X, d
     // RIMOSSA LA CUDAMEMCPY DA QUI
 }
 
-void free_device_memory(int M, int N, int k, double *Y) {
+void free_device_memory(int M, int N, int k, float *Y) {
     // IL TRASFERIMENTO (D2H) AVVIENE QUI, FUORI DAL TIMER
-    CHECK_CUDA(cudaMemcpy(Y, d_Y, (size_t)M * k * sizeof(double), cudaMemcpyDeviceToHost));
+    CHECK_CUDA(cudaMemcpy(Y, d_Y, (size_t)M * k * sizeof(float), cudaMemcpyDeviceToHost));
 
     CHECK_CUDA(cudaFree(d_A));
     CHECK_CUDA(cudaFree(d_X));
