+++ cuda/kernel_cuda_1.cu
@@ -6,12 +6,12 @@
 // =========================================================================
 // KERNEL 1: Griglia 2D + Registro Locale + Coalescing su X
 // =========================================================================
-__global__ void gemm_cuda_kernel_1(int M, int N, int k, const double *A, const double *X, double *Y) {
+__global__ void gemm_cuda_kernel_1(int M, int N, int k, const float *A, const float *X, float *Y) {
     int col = blockIdx.x * blockDim.x + threadIdx.x;
     int row = blockIdx.y * blockDim.y + threadIdx.y;
 
     if (row < M && col < k) {
-        double sum = 0.0;
+        float sum = 0.0f;
         for (int j = 0; j < N; j++) {
             sum += A[row * N + j] * X[j * k + col];
         }
@@ -22,14 +22,14 @@ __global__ void gemm_cuda_kernel_1(int M, int N, int k, const double *A, const d
 // =========================================================================
 // INTERFACCIA CPU-GPU
 // =========================================================================
-static double *d_A = NULL;
-static double *d_X = NULL;
-static double *d_Y = NULL;
+static float *d_A = NULL;
+static float*d_X = NULL;
+static float *d_Y = NULL;
 
-void setup_device_memory(int M, int N, int k, const double *A, const double *X) {
-    size_t size_A = (size_t)M * N * sizeof(double);
-    size_t size_X = (size_t)N * k * sizeof(double);
-    size_t size_Y = (size_t)M * k * sizeof(double);
+void setup_device_memory(int M, int N, int k, const float *A, const float *X) {
+    size_t size_A = (size_t)M * N * sizeof(float);
+    size_t size_X = (size_t)N * k * sizeof(float);
+    size_t size_Y = (size_t)M * k * sizeof(float);
 
     CHECK_CUDA(cudaMalloc((void **)&d_A, size_A));
     CHECK_CUDA(cudaMalloc((void **)&d_X, size_X));
@@ -39,8 +39,8 @@ void setup_device_memory(int M, int N, int k, const double *A, const double *X) 
     CHECK_CUDA(cudaMemcpy(d_X, X, size_X, cudaMemcpyHostToDevice));
 }
 
-void compute_local_gemm(int M, int N, int k, const double *A, const double *X, double *Y) {
-    CHECK_CUDA(cudaMemset(d_Y, 0, (size_t)M * k * sizeof(double)));
+void compute_local_gemm(int M, int N, int k, const float *A, const float *X, float *Y) {
+    CHECK_CUDA(cudaMemset(d_Y, 0, (size_t)M * k * sizeof(float)));
 
     dim3 threadsPerBlock(32, 8);
     dim3 blocksPerGrid(
@@ -58,9 +58,9 @@ void compute_local_gemm(int M, int N, int k, const double *A, const double *X, d
     // RIMOSSA LA CUDAMEMCPY DA QUI
 }
 
-void free_device_memory(int M, int N, int k, double *Y) {
+void free_device_memory(int M, int N, int k, float *Y) {
     // TRASFERIMENTO SPOSTATO QUI, FUORI DAL LOOP DI TIMING
-    CHECK_CUDA(cudaMemcpy(Y, d_Y, (size_t)M * k * sizeof(double), cudaMemcpyDeviceToHost));
+    CHECK_CUDA(cudaMemcpy(Y, d_Y, (size_t)M * k * sizeof(float), cudaMemcpyDeviceToHost));
 
     CHECK_CUDA(cudaFree(d_A));
     CHECK_CUDA(cudaFree(d_X));
