+++ cuda/kernel_cuda_2.cu
@@ -8,10 +8,10 @@
 // =========================================================================
 // KERNEL 3: Shared Memory Tiling + Texture/Read-Only Cache (restrict)
 // =========================================================================
-__global__ void gemm_cuda_tiled(int M, int N, int k, const double* __restrict__ A, const double* __restrict__ X, double *Y) {
+__global__ void gemm_cuda_tiled(int M, int N, int k, const float* __restrict__ A, const float* __restrict__ X, float *Y) {
 
-    __shared__ double As[TILE_SIZE][TILE_SIZE];
-    __shared__ double Xs[TILE_SIZE][TILE_SIZE];
+    __shared__ float As[TILE_SIZE][TILE_SIZE];
+    __shared__ float Xs[TILE_SIZE][TILE_SIZE];
 
     int tx = threadIdx.x;
     int ty = threadIdx.y;
@@ -19,7 +19,7 @@ __global__ void gemm_cuda_tiled(int M, int N, int k, const double* __restrict__ 
     int row = blockIdx.y * TILE_SIZE + ty;
     int col = blockIdx.x * TILE_SIZE + tx;
 
-    double sum = 0.0;
+    float sum = 0.0f;
     int numTiles = (N + TILE_SIZE - 1) / TILE_SIZE;
 
     for (int m = 0; m < numTiles; m++) {
@@ -27,13 +27,13 @@ __global__ void gemm_cuda_tiled(int M, int N, int k, const double* __restrict__ 
         if (row < M && (m * TILE_SIZE + tx) < N) {
             As[ty][tx] = __ldg(&A[row * N + (m * TILE_SIZE + tx)]);
         } else {
-            As[ty][tx] = 0.0;
+            As[ty][tx] = 0.0f;
         }
 
         if ((m * TILE_SIZE + ty) < N && col < k) {
             Xs[ty][tx] = __ldg(&X[(m * TILE_SIZE + ty) * k + col]);
         } else {
-            Xs[ty][tx] = 0.0;
+            Xs[ty][tx] = 0.0f;
         }
 
         __syncthreads();
@@ -54,14 +54,14 @@ __global__ void gemm_cuda_tiled(int M, int N, int k, const double* __restrict__ 
 // =========================================================================
 // INTERFACCIA CPU-GPU
 // =========================================================================
-static double *d_A = NULL;
-static double *d_X = NULL;
-static double *d_Y = NULL;
+static float *d_A = NULL;
+static float *d_X = NULL;
+static float *d_Y = NULL;
 
-void setup_device_memory(int M, int N, int k, const double *A, const double *X) {
-    size_t size_A = (size_t)M * N * sizeof(double);
-    size_t size_X = (size_t)N * k * sizeof(double);
-    size_t size_Y = (size_t)M * k * sizeof(double);
+void setup_device_memory(int M, int N, int k, const float *A, const float *X) {
+    size_t size_A = (size_t)M * N * sizeof(float);
+    size_t size_X = (size_t)N * k * sizeof(float);
+    size_t size_Y = (size_t)M * k * sizeof(float);
 
     CHECK_CUDA(cudaMalloc((void **)&d_A, size_A));
     CHECK_CUDA(cudaMalloc((void **)&d_X, size_X));
@@ -73,9 +73,9 @@ void setup_device_memory(int M, int N, int k, const double *A, const double *X) 
 
 // ... nel file kernel_cuda_2.cu ...
 
-void compute_local_gemm(int M, int N, int k, const double *A, const double *X, double *Y) {
+void compute_local_gemm(int M, int N, int k, const float *A, const float *X, float *Y) {
     // Rimuoviamo la cudaMemcpy da qui!
-    CHECK_CUDA(cudaMemset(d_Y, 0, (size_t)M * k * sizeof(double)));
+    CHECK_CUDA(cudaMemset(d_Y, 0, (size_t)M * k * sizeof(float)));
 
     dim3 threadsPerBlock(TILE_SIZE, TILE_SIZE);
     dim3 blocksPerGrid((k + TILE_SIZE - 1) / TILE_SIZE, (M + TILE_SIZE - 1) / TILE_SIZE);
@@ -86,9 +86,9 @@ void compute_local_gemm(int M, int N, int k, const double *A, const double *X, d
     CHECK_CUDA(cudaDeviceSynchronize());
 }
 
-void free_device_memory(int M, int N, int k, double *Y) {
+void free_device_memory(int M, int N, int k, float *Y) {
     // La copia dei dati avviene QUI, dopo che il timer nel main si è fermato
-    size_t size_Y = (size_t)M * k * sizeof(double);
+    size_t size_Y = (size_t)M * k * sizeof(float);
     CHECK_CUDA(cudaMemcpy(Y, d_Y, size_Y, cudaMemcpyDeviceToHost));
 
     CHECK_CUDA(cudaFree(d_A));
