+++ cuda/kernel_cuda_3.cu
@@ -6,7 +6,7 @@
 // =========================================================================
 // KERNEL 4: Warp-Row / Register Tiling (Zero Shared Memory Sync)
 // =========================================================================
-__global__ void gemm_warp_row_register(int M, int N, int k, const double* __restrict__ A, const double* __restrict__ X, double* Y) {
+__global__ void gemm_warp_row_register(int M, int N, int k, const float* __restrict__ A, const float* __restrict__ X, float* Y) {
     // Un warp (32 thread) si occupa di una intera riga di A.
     // blockDim.y indica quanti warp ci sono in un blocco.
     int warp_id = (blockIdx.x * blockDim.y) + threadIdx.y;
@@ -16,11 +16,11 @@ __global__ void gemm_warp_row_register(int M, int N, int k, const double* __rest
     if (warp_id < M) {
         // Se k > 32, il thread si occuperà di più colonne saltando di 32 in 32
         for (int p = lane_id; p < k; p += 32) {
-            double sum = 0.0;
+            float sum = 0.0f;
             for (int j = 0; j < N; j++) {
                 // Tutti i 32 thread del warp leggono lo STESSO elemento di A.
                 // L'hardware GPU lo trasforma in un "Broadcast" ultraveloce.
-                double a_val = A[warp_id * N + j];
+                float a_val = A[warp_id * N + j];
 
                 // Ogni thread moltiplica per la sua specifica colonna di X
                 sum += a_val * X[j * k + p];
@@ -33,14 +33,14 @@ __global__ void gemm_warp_row_register(int M, int N, int k, const double* __rest
 // =========================================================================
 // INTERFACCIA CPU-GPU
 // =========================================================================
-static double *d_A = NULL;
-static double *d_X = NULL;
-static double *d_Y = NULL;
+static float *d_A = NULL;
+static float *d_X = NULL;
+static float *d_Y = NULL;
 
-void setup_device_memory(int M, int N, int k, const double *A, const double *X) {
-    size_t size_A = (size_t)M * N * sizeof(double);
-    size_t size_X = (size_t)N * k * sizeof(double);
-    size_t size_Y = (size_t)M * k * sizeof(double);
+void setup_device_memory(int M, int N, int k, const float *A, const float *X) {
+    size_t size_A = (size_t)M * N * sizeof(float);
+    size_t size_X = (size_t)N * k * sizeof(float);
+    size_t size_Y = (size_t)M * k * sizeof(float);
 
     CHECK_CUDA(cudaMalloc((void **)&d_A, size_A));
     CHECK_CUDA(cudaMalloc((void **)&d_X, size_X));
@@ -50,8 +50,8 @@ void setup_device_memory(int M, int N, int k, const double *A, const double *X) 
     CHECK_CUDA(cudaMemcpy(d_X, X, size_X, cudaMemcpyHostToDevice));
 }
 
-void compute_local_gemm(int M, int N, int k, const double *A, const double *X, double *Y) {
-    CHECK_CUDA(cudaMemset(d_Y, 0, (size_t)M * k * sizeof(double)));
+void compute_local_gemm(int M, int N, int k, const float *A, const float *X, float *Y) {
+    CHECK_CUDA(cudaMemset(d_Y, 0, (size_t)M * k * sizeof(float)));
 
     // CONFIGURAZIONE DELLA GRIGLIA "A WARP"
     // Vogliamo blocchi 2D: X=32 (la larghezza fissa di un warp), Y=8 (8 warp per blocco)
@@ -71,13 +71,13 @@ void compute_local_gemm(int M, int N, int k, const double *A, const double *X, d
 }
 
 // ALIAS PER IL MAIN.C
-void compute_local_gemm_naive(int M, int N, int k, const double *A, const double *X, double *Y) {
+void compute_local_gemm_naive(int M, int N, int k, const float *A, const float *X, float *Y) {
     compute_local_gemm(M, N, k, A, X, Y);
 }
 
-void free_device_memory(int M, int N, int k, double *Y) {
+void free_device_memory(int M, int N, int k, float *Y) {
     // Trasferimento D2H isolato dal timer!
-    CHECK_CUDA(cudaMemcpy(Y, d_Y, (size_t)M * k * sizeof(double), cudaMemcpyDeviceToHost));
+    CHECK_CUDA(cudaMemcpy(Y, d_Y, (size_t)M * k * sizeof(float), cudaMemcpyDeviceToHost));
 
     CHECK_CUDA(cudaFree(d_A));
     CHECK_CUDA(cudaFree(d_X));
